Worked on the QML user interface

+++ src/ovito/gui/base/mainwin/ModifierListModel.h
@@ -114,10 +114,10 @@ public:
 	int listIndexFromCategoryIndex(int categoryIndex) const;
 
 	/// Returns the category index for the modifier templates.
-	int modifierTemplatesCategory() const { return _actionsPerCategory.size() - 2; }
+	int modifierTemplatesCategory() const { return (int)_actionsPerCategory.size() - 2; }
 
 	/// Returns the category index for the Python modifiers.
-	int modifierScriptsCategory() const { return _actionsPerCategory.size() - 1; }
+	int modifierScriptsCategory() const { return (int)_actionsPerCategory.size() - 1; }
 
 	/// Returns whether sorting of available modifiers into categories is enabled.
 	bool useCategories() const { return _useCategories; }
